More refactoring and integration
