+++ app/main.py
@@ -3,7 +3,7 @@ import collections
 from ftplib import Error as FTPException
 import json
 import os
-import time
+
 
 import requests
 from requests.adapters import HTTPAdapter
@@ -13,7 +13,7 @@ from sdc.crypto.decrypter import decrypt
 from sdc.crypto.exceptions import CryptoError, InvalidTokenException
 from sdc.crypto.key_store import KeyStore, validate_required_keys
 from sdc.rabbit.publisher import QueuePublisher
-from sdc.rabbit.exceptions import QuarantinableError, RetryableError, BadMessageError
+from sdc.rabbit.exceptions import QuarantinableError, RetryableError
 from sdc.rabbit.consumers import MessageConsumer
 from tornado.httpclient import AsyncHTTPClient, HTTPError
 import tornado.httpserver
@@ -23,6 +23,7 @@ import yaml
 
 from app import create_and_wrap_logger
 from app import settings
+from app.anti_virus_check import AntiVirusCheck
 from app.sdxftp import SDXFTP
 from app.settings import SERVICE_REQUEST_TOTAL_RETRIES, SERVICE_REQUEST_BACKOFF_FACTOR, RM_SDX_GATEWAY_URL
 from app.settings import BASIC_AUTH
@@ -43,7 +44,6 @@ class ConsumerError(Exception):
 
 
 Payload = collections.namedtuple('Payload', 'decoded_contents file_name case_id survey_id')
-AVResult = collections.namedtuple('AVResult', 'safe ready scan_results')
 
 
 class SeftConsumer:
@@ -111,7 +111,8 @@ class SeftConsumer:
 
             self.bound_logger.info("Retrieve results")
             if settings.ANTI_VIRUS_ENABLED:
-                self._send_for_av_scan(payload)
+                av_check = AntiVirusCheck(tx_id=tx_id)
+                av_check.send_for_av_scan(payload)
 
             file_path = self._get_ftp_file_path(payload.survey_id)
             self.bound_logger.info("Send {} to ftp server.".format(payload.file_name))
@@ -123,104 +124,6 @@ class SeftConsumer:
         except TypeError:
             self.bound_logger.exception()
             raise
-
-    def _send_for_av_scan(self, payload):
-        self.bound_logger.debug("A/V scanned enabled {}".format(settings.ANTI_VIRUS_ENABLED))
-        self.bound_logger.info("Sending for AV check")
-        data_id = self.send_for_anti_virus_check(payload.file_name, payload.decoded_contents)
-        ready = False
-        while not ready:
-            results = self.get_anti_virus_result(data_id)
-            ready = results.ready
-            if not ready:
-                time.sleep(settings.ANTI_VIRUS_WAIT_TIME)
-            elif not results.safe:
-                error = "Unsafe file detected for case id {} with filename {}".format(payload.case_id,
-                                                                                      payload.file_name)
-                self._write_scan_report(results, payload.file_name)
-                self.bound_logger.error(error)
-                raise QuarantinableError()
-
-    def add_api_key(self, headers):
-        if settings.ANTI_VIRUS_API_KEY:
-            self.bound_logger.debug("Setting A/V API key")
-            headers['apikey'] = settings.ANTI_VIRUS_API_KEY
-
-    def check_av_response(self, response):
-        if response.status_code != 200:
-            if response.status_code == 401:
-                logger.critical("Invalid OPSWAT API Key - unable to continue")
-                raise RetryableError()
-            elif 400 < response.status_code < 500:
-                raise BadMessageError()
-            elif 400 > response.status_code > 500:
-                raise RetryableError()
-        else:
-            logger.info("Response from AV OK")
-
-    def send_for_anti_virus_check(self, filename, contents):
-        url = settings.ANTI_VIRUS_BASE_URL + "file"
-        headers = {
-            "filename": filename,
-        }
-        self.add_api_key(headers)
-        response = requests.post(url=url, headers=headers, data=contents)
-
-        self.check_av_response(response)
-
-        logger.info("Response received {}".format(response.text))
-        result = response.json()
-
-        if result.get("err"):
-            logger.error("Unable to send file for anti virus scan: {}".format(result.get("err")))
-            logger.info("Waiting before attempting again")
-            time.sleep(settings.ANTI_VIRUS_WAIT_TIME)
-            logger.info("Return message to rabbit")
-            raise RetryableError()
-        else:
-            data_id = result.get("data_id")
-            logger.info("File sent successfully for anti virus scan", data_id=data_id)
-            return data_id
-
-    def get_anti_virus_result(self, data_id):
-        url = settings.ANTI_VIRUS_BASE_URL + "file/" + data_id
-        headers = {}
-        self.add_api_key(headers)
-
-        response = requests.get(url=url, headers=headers)
-
-        self.check_av_response(response)
-
-        logger.info("Response from AV {}".format(response.text))
-
-        result = response.json()
-        scan_results = result.get("scan_results")
-
-        ready = False
-        safe = False
-
-        if scan_results:
-            progress_percentage = scan_results.get("progress_percentage")
-            if int(progress_percentage) == 100:
-                ready = True
-                logger.info("Anti virus scan complete: {}".format(scan_results.get("scan_all_result_a")))
-                scan_all_result_i = scan_results.get("scan_all_result_i")
-                if int(scan_all_result_i) == 0:
-                    logger.info("File is safe")
-                    safe = True
-                else:
-                    logger.error("File is not safe")
-            else:
-                logger.info("Scan not yet complete")
-        else:
-            logger.info("Results not yet available")
-
-        return AVResult(safe=safe, ready=ready, scan_results=scan_results)
-
-    def _write_scan_report(self, av_results, filename):
-        os.makedirs("./scan_results", exist_ok=True)
-        with open("./scan_results/{}.log".format(filename), "w") as file:
-            json.dump(av_results.scan_results, file, indent=True)
 
     def _send_to_ftp(self, decoded_contents, file_path, file_name, tx_id):
         try:
